+++ src/general.rs
@@ -20,6 +20,13 @@
 //! most operations are on the order of `O(r)` as they involve looping over the
 //! variable number of indices of differently sized data blocks.
 //!
+//! # Values for r
+//! 
+//! Values for the `r` parameter can be 2 or higher, although as `r` becomes
+//! higher, more time is spent combining and rebuilding the array. Note also
+//! that when `r` equals `2`, this data structure becomes little more than a
+//! slower version of a Hashed Array Tree.
+//! 
 //! # Safety
 //!
 //! Because this data structure is allocating memory, copying bytes using raw
@@ -30,6 +37,7 @@ use super::CyclicArray;
 use std::alloc::{Layout, alloc, dealloc, handle_alloc_error};
 use std::fmt;
 use std::ops::{Index, IndexMut};
+use std::ptr::{drop_in_place, slice_from_raw_parts_mut};
 
 const USIZE_BITS: u32 = (8 * std::mem::size_of::<usize>()) as u32;
 
@@ -340,6 +348,8 @@ impl<T> OptimalArray<T> {
             self.rebuild(self.little_b + 1);
         }
         let one_b: usize = 1 << self.little_b;
+        // n.b. when r=2 the condition for combine is never met, always does a
+        // full rebuild before that happens
         if self.little_n[1] == 2 * one_b && self.little_n[0] == one_b {
             self.combine();
             // combine does nothing to the n0 block, need to fall through to the
@@ -621,8 +631,6 @@ impl<T> OptimalArray<T> {
     ///
     /// O(n) if elements are droppable, otherwise O(rN^(1/r))
     pub fn clear(&mut self) {
-        use std::ptr::{drop_in_place, slice_from_raw_parts_mut};
-
         // find the largest allocated blocks
         let mut k: usize = self.r - 1;
         while k > 1 && self.little_n[k] == 0 {
@@ -756,10 +764,11 @@ impl<T> IntoIterator for OptimalArray<T> {
         let mut me = std::mem::ManuallyDrop::new(self);
         let little_n = std::mem::take(&mut me.little_n);
         let big_a = std::mem::take(&mut me.big_a);
+        let one_b: usize = 1 << me.little_b;
         OptArrayIntoIter {
             cursor: 0,
             big_n: me.big_n,
-            little_b: me.little_b,
+            one_b,
             r: me.r,
             little_n,
             big_a,
@@ -773,9 +782,8 @@ pub struct OptArrayIntoIter<T> {
     cursor: usize,
     /// N in the paper (number of elements)
     big_n: usize,
-    /// b in the paper, the power of two that yields B, the smallest block size
-    /// in terms of elements (see page 13 of the paper)
-    little_b: usize,
+    /// smallest block size in terms of elements
+    one_b: usize,
     /// r in the paper for which the largest blocks will hold B^r elements
     r: usize,
     /// n in the paper, number of occupied data blocks for i ∈ [r - 1], and the
@@ -788,7 +796,6 @@ pub struct OptArrayIntoIter<T> {
 
 impl<T> OptArrayIntoIter<T> {
     fn locate(&self, index: usize) -> (usize, usize, usize) {
-        let one_b: usize = 1 << self.little_b;
         let mut block_size: usize = 0;
         // find index k ∈ [r−1] that contains the data block that contains
         // the element located at logical offset `index`
@@ -797,7 +804,7 @@ impl<T> OptArrayIntoIter<T> {
         while k > 0 {
             let nk = self.little_n[k];
             if nk > 0 {
-                block_size = fast_power(one_b, k as u32);
+                block_size = fast_power(self.one_b, k as u32);
                 let k_size = block_size * nk;
                 if index < (k_offset + k_size) {
                     break;
@@ -834,96 +841,63 @@ impl<T> Iterator for OptArrayIntoIter<T> {
 
 impl<T> Drop for OptArrayIntoIter<T> {
     fn drop(&mut self) {
-        use std::ptr::{drop_in_place, slice_from_raw_parts_mut};
-
-        // find the largest allocated blocks
-        let mut k: usize = self.r - 1;
-        while k > 0 && self.little_n[k] == 0 {
-            k -= 1;
-        }
-        let one_b: usize = 1 << self.little_b;
-
-        if self.big_n > 0 && std::mem::needs_drop::<T>() {
-            // drop items and deallocate the data blocks
-            let (first_level, first_block, first_slot) = self.locate(self.cursor);
-            let (last_level, last_block, last_slot) = self.locate(self.big_n - 1);
-            if first_level == last_level && first_block == last_block {
-                let len = fast_power(one_b, first_level as u32);
-                let ptr = self.big_a[first_level][first_block];
-                // special-case, remaining values are in only one block
-                if first_slot <= last_slot {
+        // all visited elements have already been dropped, so carefully examine
+        // all of the blocks and determine which ones have unvisited elements
+        let (first_level, first_block, first_slot) = self.locate(self.cursor);
+        let (last_level, last_block, last_slot) = if self.big_n == 0 {
+            self.locate(0)
+        } else {
+            self.locate(self.big_n - 1)
+        };
+        let not_finished = self.cursor < self.big_n;
+        for level in 1..self.r {
+            let block_len = fast_power(self.one_b, level as u32);
+            let layout = Layout::array::<T>(block_len).expect("unexpected overflow");
+            let mut block = 0;
+            while let Some(ptr) = self.big_a[level].pop_front() {
+                let unvisited = (level != first_level || block >= first_block)
+                    && (level != last_level || block < last_block);
+                let is_first = level == first_level && block == first_block;
+                let is_last = level == last_level && block == last_block;
+                if is_first && is_last {
+                    // special-case for first block being the last block
+                    if first_slot <= last_slot {
+                        unsafe {
+                            drop_in_place(slice_from_raw_parts_mut(
+                                ptr.add(first_slot),
+                                last_slot - first_slot + 1,
+                            ));
+                        }
+                    }
+                } else if is_first {
+                    // special-case for partially visited first block
                     unsafe {
-                        // last_slot is pointing at the last element, need to
-                        // add one to include it in the slice
                         drop_in_place(slice_from_raw_parts_mut(
                             ptr.add(first_slot),
-                            last_slot - first_slot + 1,
+                            block_len - first_slot,
                         ));
                     }
+                } else if is_last && not_finished {
+                    // special-case for maybe partially filled last block
+                    unsafe {
+                        drop_in_place(slice_from_raw_parts_mut(ptr, last_slot + 1));
+                    }
+                } else if unvisited {
+                    // all other unvisited blocks are dropped in their entirety
+                    unsafe {
+                        drop_in_place(slice_from_raw_parts_mut(ptr, block_len));
+                    }
                 }
-                let layout = Layout::array::<T>(len).expect("unexpected overflow");
                 unsafe {
                     dealloc(ptr as *mut u8, layout);
                 }
-            } else {
-                // first partial block needs special care
-                let block_len = fast_power(one_b, first_level as u32);
-                unsafe {
-                    drop_in_place(slice_from_raw_parts_mut(
-                        self.big_a[first_level][first_block].add(first_slot),
-                        block_len - first_slot,
-                    ));
-                }
-
-                // deallocate all blocks already visited
-                for _ in 0..=first_block {
-                    let ptr = self.big_a[first_level].pop_front().unwrap();
-                    let layout = Layout::array::<T>(block_len).expect("unexpected overflow");
-                    unsafe { dealloc(ptr as *mut u8, layout) }
-                }
-
-                // smallest block needs special care
-                if self.little_n[0] > 0 {
-                    let ptr = self.big_a[1].pop_back().unwrap();
-                    unsafe {
-                        drop_in_place(slice_from_raw_parts_mut(ptr, self.little_n[0]));
-                        let layout = Layout::array::<T>(one_b).expect("unexpected overflow");
-                        dealloc(ptr as *mut u8, layout);
-                    }
-                }
-
-                // drop all elements in all remaining blocks
-                for i in (1..=k).rev() {
-                    let len = fast_power(one_b, i as u32);
-                    while let Some(ptr) = self.big_a[i].pop_front() {
-                        unsafe {
-                            drop_in_place(slice_from_raw_parts_mut(ptr, len));
-                            let layout = Layout::array::<T>(len).expect("unexpected overflow");
-                            dealloc(ptr as *mut u8, layout);
-                        }
-                    }
-                }
-            }
-        } else {
-            // no drop, just deallocate the data blocks
-            for i in (1..=k).rev() {
-                let len = fast_power(one_b, i as u32);
-                while let Some(ptr) = self.big_a[i].pop_front() {
-                    unsafe {
-                        let layout = Layout::array::<T>(len).expect("unexpected overflow");
-                        dealloc(ptr as *mut u8, layout);
-                    }
-                }
+                block += 1;
             }
         }
 
         // zero out everything to the initial state
         self.cursor = 0;
         self.big_n = 0;
-        self.little_b = 2;
-        for idx in 0..self.little_n.len() {
-            self.little_n[idx] = 0;
-        }
     }
 }
 
@@ -1013,34 +987,57 @@ mod tests {
 
     #[test]
     fn test_general_array_push_many_ints_r2() {
-        // TODO: leaks memory (64 bytes) for each push/pop cycle
         let mut sut: OptimalArray<usize> = OptimalArray::with_r(2);
         for value in 0..1_000_000 {
             sut.push(value);
         }
         // Rebuild(2B) will have been called 9 times
         assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1000448);
         for index in 0..1_000_000 {
             assert_eq!(sut[index], index);
         }
-        assert_eq!(sut[99_999], 99_999);
-        assert_eq!(sut.capacity(), 1000448);
-
+        // TODO: leaks memory (64 bytes) for each pop cycle; it is a data block
+        // of 8 elements, likely when going from b=4 to b=3, or b=3 to b=2
         for value in (0..1_000_000).rev() {
             assert_eq!(sut.pop(), Some(value));
         }
         assert_eq!(sut.len(), 0);
         assert_eq!(sut.capacity(), 0);
+
+        // for each push/pop iteration:
+        //
+        // empty:   OptimalArray(n: 0,       l: 0,     h: 16,      b: 2,  r: 2, n: 0,   0)
+        // rebuild: OptimalArray(n: 16,      l: 0,     h: 16,      b: 2,  r: 2, n: 4,   4)
+        // rebuild: OptimalArray(n: 64,      l: 4,     h: 64,      b: 3,  r: 2, n: 8,   8)
+        // rebuild: OptimalArray(n: 256,     l: 16,    h: 256,     b: 4,  r: 2, n: 16,  16)
+        // rebuild: OptimalArray(n: 1024,    l: 64,    h: 1024,    b: 5,  r: 2, n: 32,  32)
+        // rebuild: OptimalArray(n: 4096,    l: 256,   h: 4096,    b: 6,  r: 2, n: 64,  64)
+        // rebuild: OptimalArray(n: 16384,   l: 1024,  h: 16384,   b: 7,  r: 2, n: 128, 128)
+        // rebuild: OptimalArray(n: 65536,   l: 4096,  h: 65536,   b: 8,  r: 2, n: 256, 256)
+        // rebuild: OptimalArray(n: 262144,  l: 16384, h: 262144,  b: 9,  r: 2, n: 512, 512)
+        // testing: OptimalArray(n: 1000000, l: 65536, h: 1048576, b: 10, r: 2, n: 576, 977)
+        // rebuild: OptimalArray(n: 65536,   l: 65536, h: 1048576, b: 10, r: 2, n: 1024,64)
+        // rebuild: OptimalArray(n: 16384,   l: 16384, h: 262144,  b: 9,  r: 2, n: 512, 32)
+        // rebuild: OptimalArray(n: 4096,    l: 4096,  h: 65536,   b: 8,  r: 2, n: 256, 16)
+        // rebuild: OptimalArray(n: 1024,    l: 1024,  h: 16384,   b: 7,  r: 2, n: 128, 8)
+        // rebuild: OptimalArray(n: 256,     l: 256,   h: 4096,    b: 6,  r: 2, n: 64,  4)
+        // rebuild: OptimalArray(n: 64,      l: 64,    h: 1024,    b: 5,  r: 2, n: 32,  2)
+        // rebuild: OptimalArray(n: 16,      l: 16,    h: 256,     b: 4,  r: 2, n: 16,  1)
+        // rebuild: OptimalArray(n: 4,       l: 4,     h: 64,      b: 3,  r: 2, n: 4,   1)
+        // empty:   OptimalArray(n: 0,       l: 0,     h: 16,      b: 2,  r: 2, n: 0,   0)
 
         // and do it again to be sure shrinking works correctly
-        for value in 0..1_000_000 {
-            sut.push(value);
-        }
-        for value in (0..1_000_000).rev() {
-            assert_eq!(sut.pop(), Some(value));
-        }
-        assert_eq!(sut.len(), 0);
-        assert_eq!(sut.capacity(), 0);
+        // for value in 0..1_000_000 {
+        //     sut.push(value);
+        // }
+        // assert_eq!(sut.len(), 1_000_000);
+        // assert_eq!(sut.capacity(), 1000448);
+        // for value in (0..1_000_000).rev() {
+        //     assert_eq!(sut.pop(), Some(value));
+        // }
+        // assert_eq!(sut.len(), 0);
+        // assert_eq!(sut.capacity(), 0);
     }
 
     #[test]
@@ -1131,6 +1128,122 @@ mod tests {
         }
         for value in (0..1_000_000).rev() {
             assert_eq!(sut.pop(), Some(value));
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+    }
+
+    #[test]
+    fn test_general_array_push_many_strings_r2() {
+        let mut sut = OptimalArray::<String>::with_r(2);
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1000448);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+
+        // and do it again to be sure shrinking works correctly
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1000448);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+    }
+
+    #[test]
+    fn test_general_array_push_many_strings_r3() {
+        let mut sut = OptimalArray::<String>::with_r(3);
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_064);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+
+        // and do it again to be sure shrinking works correctly
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_064);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+    }
+
+    #[test]
+    fn test_general_array_push_many_strings_r4() {
+        let mut sut = OptimalArray::<String>::with_r(4);
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_000);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+
+        // and do it again to be sure shrinking works correctly
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_000);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+    }
+
+    #[test]
+    fn test_general_array_push_many_strings_r5() {
+        let mut sut = OptimalArray::<String>::with_r(5);
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_000);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
+        }
+        assert_eq!(sut.len(), 0);
+        assert_eq!(sut.capacity(), 0);
+
+        // and do it again to be sure shrinking works correctly
+        for _ in 0..1_000_000 {
+            let value = ulid::Ulid::new().to_string();
+            sut.push(value);
+        }
+        assert_eq!(sut.len(), 1_000_000);
+        assert_eq!(sut.capacity(), 1_000_000);
+        while let Some(value) = sut.pop() {
+            assert_eq!(value.len(), 26);
         }
         assert_eq!(sut.len(), 0);
         assert_eq!(sut.capacity(), 0);
@@ -1283,6 +1396,128 @@ mod tests {
                 sut.push(value);
             }
             assert_eq!(sut.len(), 1_000);
+        }
+    }
+
+    #[test]
+    fn test_general_array_into_iterator_edge_case() {
+        // add 4, iterate 3
+        let inputs = ["one", "two", "three", "four"];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..3 {
+            iter.next();
+        }
+
+        // add 4, iterate 4
+        let inputs = ["one", "two", "three", "four"];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        assert_eq!(4, sut.into_iter().count());
+
+        // add 8, iterate 3
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..3 {
+            iter.next();
+        }
+
+        // add 8, iterate 5
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..5 {
+            iter.next();
+        }
+
+        // add 8, iterate 8
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        assert_eq!(8, sut.into_iter().count());
+
+        // add 7, iterate 3
+        let inputs = ["one", "two", "three", "four", "five", "six", "seven"];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..3 {
+            iter.next();
+        }
+
+        // add 7, iterate 5
+        let inputs = ["one", "two", "three", "four", "five", "six", "seven"];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..5 {
+            iter.next();
+        }
+
+        // add 13, iterate 3
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten",
+            "eleven", "twelve", "thirteen",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..3 {
+            iter.next();
+        }
+
+        // add 13, iterate 5
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten",
+            "eleven", "twelve", "thirteen",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..5 {
+            iter.next();
+        }
+
+        // add 13, iterate 9
+        let inputs = [
+            "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten",
+            "eleven", "twelve", "thirteen",
+        ];
+        let mut sut: OptimalArray<String> = OptimalArray::new();
+        for item in inputs {
+            sut.push(item.to_owned());
+        }
+        let mut iter = sut.into_iter();
+        for _ in 0..9 {
+            iter.next();
         }
     }