+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Fixed
+- Memory leak in simple implementation.
+- Memory leaks in general and simple `IntoIterator` implementations.
+
 ## [1.0.1] - 2025-09-24
 ### Changed
 - Improve performance of `pop()` and `swap_remove()` in some cases.